+++ nestsClient/src/commonMain/kotlin/com/vitorpamplona/nestsclient/moq/lite/MoqLiteSession.kt
@@ -34,7 +34,6 @@ import kotlinx.coroutines.Job
 import kotlinx.coroutines.cancelAndJoin
 import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.consumeAsFlow
 import kotlinx.coroutines.launch
@@ -1370,157 +1369,4 @@ class MoqLiteSession internal constructor(
             pumpScope: CoroutineScope,
         ): MoqLiteSession = MoqLiteSession(transport, pumpScope)
     }
-}
-
-/**
- * One frame received from a subscription. moq-lite's wire format
- * carries no per-frame envelope beyond the size; [groupSequence] is
- * pulled from the group header so consumers can detect group rollover
- * (e.g. for keyframe boundaries).
- */
-data class MoqLiteFrame(
-    val groupSequence: Long,
-    val payload: ByteArray,
-) {
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (other !is MoqLiteFrame) return false
-        return groupSequence == other.groupSequence && payload.contentEquals(other.payload)
-    }
-
-    override fun hashCode(): Int = 31 * groupSequence.hashCode() + payload.contentHashCode()
-}
-
-/**
- * Active subscription handle returned by [MoqLiteSession.subscribe].
- * [frames] emits every frame the publisher pushes; [unsubscribe]
- * FINs the bidi to signal "no longer interested" (moq-lite has no
- * UNSUBSCRIBE message — FIN is the protocol).
- */
-class MoqLiteSubscribeHandle internal constructor(
-    val id: Long,
-    val ok: MoqLiteSubscribeOk,
-    val frames: Flow<MoqLiteFrame>,
-    private val unsubscribeAction: suspend () -> Unit,
-) {
-    suspend fun unsubscribe() = unsubscribeAction()
-}
-
-/**
- * Active announce-discovery handle returned by [MoqLiteSession.announce].
- * [updates] emits every [MoqLiteAnnounce] update the relay streams
- * back; [close] FINs the bidi to stop receiving updates.
- */
-class MoqLiteAnnouncesHandle internal constructor(
-    val updates: Flow<MoqLiteAnnounce>,
-    private val close: suspend () -> Unit,
-) {
-    suspend fun close() = close.invoke()
-}
-
-/** Thrown when subscribe is rejected (Drop) or the response stream dies. */
-class MoqLiteSubscribeException(
-    message: String,
-    cause: Throwable? = null,
-) : RuntimeException(message, cause)
-
-/**
- * Active publisher handle returned by [MoqLiteSession.publish].
- *
- * Lifecycle:
- *   1. Call [startGroup] (or [send] which auto-starts a fresh group on
- *      first call) to begin pushing frames for one Opus group.
- *   2. Call [send] for each frame (one Opus packet = one frame).
- *   3. Call [endGroup] to FIN the current group's uni stream and start
- *      a fresh group on the next [send]. Group rollover is the
- *      publisher's call — typically every N seconds or every keyframe.
- *   4. Call [close] when the broadcast ends — sends `Announce(Ended)`
- *      on every active announce bidi and FINs every group stream.
- */
-interface MoqLitePublisherHandle {
-    /**
-     * The broadcast suffix this publisher claimed at [MoqLiteSession.publish].
-     * Always normalised per [MoqLitePath].
-     */
-    val suffix: String
-
-    /**
-     * The next group sequence number that will be assigned by [send] /
-     * [startGroup]. Snapshot-only — read AFTER the broadcaster has
-     * stopped sending into this publisher (typically just before the
-     * caller closes the publisher in a hot-swap), so the value is the
-     * highest-already-used sequence + 1.
-     *
-     * Used by [com.vitorpamplona.nestsclient.MoqLiteNestsSpeaker]'s
-     * hot-swap path to seed the new session's publisher with a
-     * monotonically-continuing sequence — without this, every JWT
-     * refresh restarts at sequence 0 and kixelated/hang's
-     * `Container.Consumer.#run` drops every group whose sequence is
-     * less than its current `#active` high-water mark, killing audio
-     * for the watcher until either `#active` rolls over or the
-     * watcher re-subscribes.
-     *
-     * `@Volatile` on the implementation; safe to read from any
-     * coroutine. The accept-tiny-race window between read and a
-     * concurrent `send` is closed in practice because the broadcaster
-     * is responsible for swapping its publisher reference BEFORE the
-     * caller reads this value (so no further sends land on this
-     * publisher).
-     */
-    val nextSequence: Long
-
-    /**
-     * Start a new group. Allocates a fresh sequence id and opens a new
-     * uni stream pre-loaded with `DataType=Group + GroupHeader`. Idempotent
-     * — calling [startGroup] when the previous group hasn't been ended
-     * is treated as an implicit [endGroup] then a new start.
-     */
-    suspend fun startGroup()
-
-    /**
-     * Push one [payload] (one Opus packet) as a `varint(size) + payload`
-     * frame on the current group's uni stream. Auto-starts a group if
-     * none is active.
-     *
-     * Returns false if no inbound subscriber is currently attached.
-     * Subscriber-less sends silently drop on the wire — the relay keeps
-     * the publisher's announce active either way, so unmute is
-     * sample-accurate.
-     */
-    suspend fun send(payload: ByteArray): Boolean
-
-    /** FIN the current group's uni stream. The next [send] starts a fresh group. */
-    suspend fun endGroup()
-
-    /**
-     * Register a callback that fires once each time a new inbound
-     * subscriber is registered against this publisher's track (i.e.
-     * each track-matching SUBSCRIBE bidi the relay opens to us). Used
-     * to push a "track-latest" payload — the canonical example is the
-     * broadcast catalog manifest, which a watcher needs to receive on
-     * subscribe but doesn't change between subscribers — without
-     * forcing the publisher to maintain a periodic re-emit loop.
-     *
-     * Called once per accepted SUBSCRIBE (track filter passed). Fires
-     * OUTSIDE the publisher's serialisation lock, so the hook can
-     * safely call [send] / [endGroup] without deadlocking.
-     *
-     * Caller MUST set the hook before any subscriber attaches (typically
-     * immediately after [com.vitorpamplona.nestsclient.moq.lite.MoqLiteSession.publish]
-     * returns) — there's no "fire-on-set for existing subscribers"
-     * replay. For the typical catalog use case the publisher is fresh
-     * when the hook is set, and the relay's SUBSCRIBE bidi takes a
-     * round-trip to arrive, so this is safe in practice.
-     *
-     * Pass `null` to clear the hook. Calling twice with non-null
-     * replaces the previous hook (no de-duplication).
-     */
-    fun setOnNewSubscriber(hook: (suspend () -> Unit)?)
-
-    /**
-     * Stop publishing. Sends `Announce(Ended)` on every active announce
-     * bidi, FINs the current group, and releases all per-publisher
-     * resources. Idempotent.
-     */
-    suspend fun close()
 }
